Align with AUTH48 changes (aka RFC9882)
The only differences will be the refs:

1. Sean and I don't know how to make it refer to:
https://www.rfc-editor.org/info/
instead of https://www.rfc-editor.org/rfc/
for the RFC references.

I think maybe the DOI dereference from the tools looks slightly different than the format from the RFC editor.

In both instances it does not matter because they point to the same URI.